Initial commit.
Using XAMPP to host a PHP webserver. This repository resides in the
XAMPP installation folder.
All the original XAMPP files have been ignored from this repository.
Precisely, every file that is not in one of the following folders is
excluded from the repository:
- xampp/messaging_db
- xampp/messaging_docs
- xampp/htdocs/messaging

All the files under 'xampp/htdocs' are accessible via the webserver,
everything else is not.

The current directory tree structure is subject to change. This is just
a very raw start.